+++ client/src/App.tsx
@@ -10,11 +10,13 @@ import CheckPassword from "./page/User/CheckPassword";
 import ProfileUpdate from "./page/User/ProfileUpdate";
 import clsx from "clsx";
 import { useErrorModal } from "./state/errorModalStore";
-import { useLayoutEffect } from "react";
+import { useEffect, useLayoutEffect } from "react";
 import ErrorModal from "./component/utils/ErrorModal";
 import ChatTestPage from "./page/Chat/ChatTestPage";
 import ChatRoom from "./component/Chats/ChatRoom/ChatRoom";
 import ChatRooms from "./component/Chats/ChatRooms/ChatRooms";
+import { useUserStore } from "./state/store";
+import { io, Socket } from "socket.io-client";
 
 function MainContainer({ children }: { children: React.ReactNode }) {
 	const location = useLocation();
@@ -46,6 +48,32 @@ function App() {
 	useLayoutEffect(() => {
 		errorModal.clear();
 	}, []);
+
+	const isLogin = useUserStore(state => state.isLogin);
+
+	useEffect(() => {
+		if (isLogin) {
+			// 로그인 성공 시 chat_server에 소켓 연결
+			const socket: Socket = io(
+				`${import.meta.env.VITE_CHAT_ADDRESS}/chat`
+			);
+
+			socket.on("connect", () => {
+				console.log(
+					"Connected to chat server with socket ID:",
+					socket.id
+				);
+			});
+
+			socket.on("disconnect", () => {
+				console.log("Disconnected from chat server");
+			});
+
+			socket.on("update_user_list", userList => {
+				console.log("현재 접속 중인 사용자 목록:", userList);
+			});
+		}
+	}, [isLogin]);
 
 	return (
 		<div className={AppContainer}>
